+++ src/mjlab/tasks/tennis/config/g1/__init__.py
@@ -10,6 +10,8 @@ from .env_cfgs import (
   unitree_g1_tennis_latent_cross_env_cfg,
   unitree_g1_tennis_latent_hit_env_cfg,
   unitree_g1_tennis_sonic_cross_env_cfg,
+  unitree_g1_tennis_sonic_encoder_cross_env_cfg,
+  unitree_g1_tennis_sonic_encoder_hit_env_cfg,
   unitree_g1_tennis_sonic_hit_env_cfg,
 )
 from .rl_cfg import (
@@ -20,6 +22,8 @@ from .rl_cfg import (
   unitree_g1_tennis_latent_cross_ppo_runner_cfg,
   unitree_g1_tennis_latent_ppo_runner_cfg,
   unitree_g1_tennis_sonic_cross_ppo_runner_cfg,
+  unitree_g1_tennis_sonic_encoder_cross_ppo_runner_cfg,
+  unitree_g1_tennis_sonic_encoder_hit_ppo_runner_cfg,
   unitree_g1_tennis_sonic_hit_ppo_runner_cfg,
 )
 
@@ -94,11 +98,31 @@ register_mjlab_task(
 )
 
 register_mjlab_task(
+  task_id="Mjlab-Tennis-Hit-SONIC-Encoder-Unitree-G1",
+  env_cfg=unitree_g1_tennis_sonic_encoder_hit_env_cfg(court_size=DEFAULT_COURT_SIZE),
+  play_env_cfg=unitree_g1_tennis_sonic_encoder_hit_env_cfg(
+    play=True, court_size=DEFAULT_COURT_SIZE
+  ),
+  rl_cfg=unitree_g1_tennis_sonic_encoder_hit_ppo_runner_cfg(),
+  runner_cls=TennisTokenOnPolicyRunner,
+)
+
+register_mjlab_task(
   task_id="Mjlab-Tennis-Cross-SONIC-Unitree-G1",
   env_cfg=unitree_g1_tennis_sonic_cross_env_cfg(court_size=DEFAULT_COURT_SIZE),
   play_env_cfg=unitree_g1_tennis_sonic_cross_env_cfg(
     play=True, court_size=DEFAULT_COURT_SIZE
   ),
   rl_cfg=unitree_g1_tennis_sonic_cross_ppo_runner_cfg(),
+  runner_cls=TennisTokenOnPolicyRunner,
+)
+
+register_mjlab_task(
+  task_id="Mjlab-Tennis-Cross-SONIC-Encoder-Unitree-G1",
+  env_cfg=unitree_g1_tennis_sonic_encoder_cross_env_cfg(court_size=DEFAULT_COURT_SIZE),
+  play_env_cfg=unitree_g1_tennis_sonic_encoder_cross_env_cfg(
+    play=True, court_size=DEFAULT_COURT_SIZE
+  ),
+  rl_cfg=unitree_g1_tennis_sonic_encoder_cross_ppo_runner_cfg(),
   runner_cls=TennisTokenOnPolicyRunner,
 )